修正由使用者發文時建立的子分類，該使用者群組無權限在裡面發文的bug。
1.修正自訂頁面區塊若沒資料就不要秀出空區塊
2.修正自訂頁面區塊若沒有指定目錄，則自動找一個自訂頁面目錄來呈現
3.修正由使用者發文時建立的子分類，該使用者群組無權限在裡面發文的bug。
4.調整檔案上傳欄位col_sn為mediumint(9)格式
5.調整uid欄位為mediumint格式
6.修正子分類若轉為自訂頁面分類或新聞分類時有時分類會消失的問題。
7.修正點選公告人不會列出該公告者相關新聞的bug

+++ sql/mysql.english.sql
@@ -1,7 +1,7 @@
 CREATE TABLE `tadnews_files_center` (
   `files_sn` smallint(5) unsigned NOT NULL AUTO_INCREMENT,
   `col_name` varchar(255) NOT NULL default '',
-  `col_sn` smallint(5) unsigned NOT NULL default 0,
+  `col_sn` mediumint(9) unsigned NOT NULL default 0,
   `sort` smallint(5) unsigned NOT NULL default 0,
   `kind` enum('img','file') NOT NULL default 'img',
   `file_name` varchar(255) NOT NULL default '',
